+++ examples/basic/src/model-function/generate-speech-dynamic-model.ts
@@ -1,23 +1,23 @@
 import dotenv from "dotenv";
 import {
-  ElevenLabsSpeechSynthesisModel,
-  LmntSpeechSynthesisModel,
-  SpeechSynthesisModel,
-  synthesizeSpeech,
+  ElevenLabsSpeechModel,
+  LmntSpeechModel,
+  SpeechGenerationModel,
+  generateSpeech,
 } from "modelfusion";
 
 dotenv.config();
 
-async function callModel(model: SpeechSynthesisModel) {
-  return await synthesizeSpeech(model, "Hello, World!");
+async function callModel(model: SpeechGenerationModel) {
+  return await generateSpeech(model, "Hello, World!");
 }
 async function main() {
   const model =
     Math.random() < 0.5
-      ? new LmntSpeechSynthesisModel({
+      ? new LmntSpeechModel({
           voice: "034b632b-df71-46c8-b440-86a42ffc3cf3", // Henry
         })
-      : new ElevenLabsSpeechSynthesisModel({
+      : new ElevenLabsSpeechModel({
           voice: "pNInz6obpgDQGcFmaJgB", // Adam
         });