+++ examples/basic/src/model-provider/lmnt/lmnt-generate-speech-example.ts
@@ -1,12 +1,12 @@
 import dotenv from "dotenv";
-import { LmntSpeechSynthesisModel, synthesizeSpeech } from "modelfusion";
+import { LmntSpeechModel, generateSpeech } from "modelfusion";
 import fs from "node:fs";
 
 dotenv.config();
 
 async function main() {
-  const speech = await synthesizeSpeech(
-    new LmntSpeechSynthesisModel({
+  const speech = await generateSpeech(
+    new LmntSpeechModel({
       voice: "034b632b-df71-46c8-b440-86a42ffc3cf3", // Henry
     }),
     "Good evening, ladies and gentlemen! Exciting news on the airwaves tonight " +