+++ src/model-provider/lmnt/index.ts
@@ -1,2 +1,2 @@
 export * from "./LmntApiConfiguration.js";
-export * from "./LmntSpeechSynthesisModel.js";
+export * from "./LmntSpeechModel.js";